ai-engine/requirements.txt
fastapi
uvicorn
sentence-transformers
numpy
faiss-cpu
pydantic